backendPy/requirements.txt
# ---- Core API ----
fastapi==0.110.0
uvicorn[standard]==0.27.1

# ---- Auth & Security ----
python-jose==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

# ---- Database ----
pymongo==4.6.1
motor==3.3.2

# ---- Config & Env ----
python-dotenv==1.0.1
pydantic==2.6.1
pydantic-settings==2.2.1
email-validator==2.1.1
python-multipart==0.0.9

# ---- HTTP ----
requests
httpx==0.27.0
orjson==3.10.3

# ---- ML / RAG stack ----
torch
transformers
accelerate
sentence-transformers
faiss-cpu
pdfplumber
python-docx
langchain
langchain-community
langchain-text-splitters
bitsandbytes
datasets
numpy
